Fix unsafe code in 'caveman.nix'; fix obsidian-missing skipping downloading files besides README.md

+++ darwin/packages/codex/extensions/caveman.nix
@@ -88,13 +88,11 @@ let
 
         printf 'Installing Caveman skill: %s\n' "$skill_name"
 
-        rm -rf -- "$shared_skills/$skill_name"
-
+        rm -rf -- "${shared_skills:?}/${skill_name:?}"
+        
         cp -R \
           "$skill" \
-          "$shared_skills/$skill_name"
-      done
-
+          "${shared_skills:?}/${skill_name:?}"
 
       # COMPLETE
       # =========================
